llm_convo_signalwire/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# SignalWire
signalwire==2.1.1

# AI Services
groq>=0.9.0
deepgram-sdk>=3.4.0
boto3==1.34.0  # for Amazon Polly
googlemaps==4.10.0

# Data and Caching
redis==5.0.1
pydantic==2.5.0
sqlalchemy==2.0.23

# Audio Processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles>=23.2.1
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0 